input, aritmatika volume balok

+++ Day099.dart
@@ -2,15 +2,15 @@ import 'dart:io';
 
 void main() {
   stdout.write("Masukkan nilai panjang : ");
-  int panjang = int.parse(stdin.readLineSync()!);
+  double panjang = double.parse(stdin.readLineSync()!);
 
   stdout.write("Masukkan nilai lebar : ");
-  int lebar = int.parse(stdin.readLineSync()!);
+  double lebar = double.parse(stdin.readLineSync()!);
 
   stdout.write("Masukkan nilai tinggi : ");
-  int tinggi = int.parse(stdin.readLineSync()!);
+  double tinggi = double.parse(stdin.readLineSync()!);
 
-  int volumeBalok = panjang * lebar * tinggi;
+  double volumeBalok = panjang * lebar * tinggi;
 
   print("hasil dari rumus $panjang x $lebar x $tinggi = $volumeBalok");
 }
